+++ src/main/scala/engine/core/render_element/RenderElement.scala
@@ -1,0 +1,7 @@
+package engine.core.render_element
+
+import engine.render.renderer.render_data.RenderData
+
+trait RenderElement {
+  def renderData: RenderData
+}
